+++ NppCSharpPluginPack/Properties/AssemblyInfo.cs
@@ -28,5 +28,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("0.0.2.3")]
-[assembly: AssemblyFileVersion("0.0.2.3")]
+[assembly: AssemblyVersion("0.0.2.4")]
+[assembly: AssemblyFileVersion("0.0.2.4")]